moved some llm service code into helper file, tweaked political prompt to bypass historian as well
